[rocprofiler-sdk] Introduce TopologyProvider abstraction
Factor GPU/CPU agent discovery behind a polymorphic TopologyProvider
interface so the existing Linux/KFD path is preserved verbatim while
WSL and native Windows gain dedicated backends.

  - topology/linux_kfd_provider: byte-identical move of read_topology()
    and its helpers (parse_cpu_info, read_map, read_property, etc.) out
    of agent.cpp. Linux behaviour is unchanged.
  - topology/dxg_provider: WSL backend skeleton. Dlopens libdxcore.so
    and probes CreateDXCoreAdapterFactory; full DXCore-to-agent
    enumeration is left as a follow-up.
  - topology/d3dkmt_provider: native Windows backend skeleton. All
    D3DKMT calls are #ifdef _WIN32 guarded so the class still links on
    Linux (stubs return empty + warning). The full enumeration body
    is left as a follow-up.
  - topology/topology_provider: factory make_default_provider() with
    runtime selection precedence:
      ROCPROFILER_FORCE_DXG -> ROCPROFILER_FORCE_D3DKMT -> sysfs ->
      libdxcore.so loadable -> _WIN32 -> LinuxKfd fallback.
    The ROCPROFILER_FORCE_* env vars are new and intended for triage.
  - agent.cpp: read_topology() and its helpers deleted. The singleton
    accessor and internal_refresh_topology() now go through the
    provider. No other public function changed.
  - agent_internal.hpp: minimal forward declaration of
    update_agent_runtime_visibility so providers can call it without
    touching the public header.
