+++ projects/rocprofiler-sdk/source/lib/rocprofiler-sdk/CMakeLists.txt
@@ -56,6 +56,7 @@ add_subdirectory(rccl)
 add_subdirectory(rocdecode)
 add_subdirectory(rocjpeg)
 add_subdirectory(details)
+add_subdirectory(topology)
 add_subdirectory(ompt)
 add_subdirectory(registration)
